made it so that the ssh agent is only started on a darwin computer

+++ .zshrc
@@ -98,32 +98,33 @@ if [ ! -z $VLESS ]; then
 fi
 
 # Start/Reuse SSH Agent - restart or re-use an existing agent
-SSH_AGENT_CACHE=/tmp/ssh_agent_eval_`whoami`
-if [ -s "${SSH_AGENT_CACHE}" ]
-then
-    echo "Reusing existing ssh-agent"
-    eval `cat "${SSH_AGENT_CACHE}"`
-# Check that agent still exists
-    kill -0 "${SSH_AGENT_PID}" 2>/dev/null
-    if [ $? -eq 1 ]
+if [[ "$OSTYPE" =~ ^darwin ]]; then
+    SSH_AGENT_CACHE=/tmp/ssh_agent_eval_`whoami`
+    if [ -s "${SSH_AGENT_CACHE}" ]
     then
-    echo "ssh-agent pid ${SSH_AGENT_PID} no longer running"
-        # Looks like the SSH-Agent has died, it'll be restarted below
-        rm -f "${SSH_AGENT_CACHE}"
+        echo "Reusing existing ssh-agent"
+        eval `cat "${SSH_AGENT_CACHE}"`
+        # Check that agent still exists
+        kill -0 "${SSH_AGENT_PID}" 2>/dev/null
+        if [ $? -eq 1 ]
+        then
+            echo "ssh-agent pid ${SSH_AGENT_PID} no longer running"
+            # Looks like the SSH-Agent has died, it'll be restarted below
+            rm -f "${SSH_AGENT_CACHE}"
+        fi
+    fi
+
+    if [ ! -f "${SSH_AGENT_CACHE}" ]
+    then
+        echo "Starting new ssh-agent"
+        touch "${SSH_AGENT_CACHE}"
+        chmod 600 "${SSH_AGENT_CACHE}"
+        ssh-agent >> "${SSH_AGENT_CACHE}"
+        chmod 400 "${SSH_AGENT_CACHE}"
+        eval `cat "${SSH_AGENT_CACHE}"`
+        ssh-add ~/.ssh/id_rsa
     fi
 fi
-
-if [ ! -f "${SSH_AGENT_CACHE}" ]
-then
-    echo "Starting new ssh-agent"
-    touch "${SSH_AGENT_CACHE}"
-    chmod 600 "${SSH_AGENT_CACHE}"
-    ssh-agent >> "${SSH_AGENT_CACHE}"
-    chmod 400 "${SSH_AGENT_CACHE}"
-    eval `cat "${SSH_AGENT_CACHE}"`
-    ssh-add ~/.ssh/id_rsa
-fi
-
 # Load my aliases/functions:
 . ~/.aliases
 . ~/.functions
